splitting responsibilities to BuildFileModifier classes
