+++ lts-test/src/test/resources/repositories.yaml
@@ -27,7 +27,7 @@ repositories:
           :sdks:java:io:google-cloud-platform:check \
           :sdks:java:extensions:google-cloud-platform-core:check \
           :runners:google-cloud-dataflow-java:check \
-           -x analyzeClassesDependencies
+          -x analyzeClassesDependencies
   - name: google-api-java-client
     url: https://github.com/googleapis/google-api-java-client.git
     tag: v1.31.5
